Move auth logic to service
