+++ src/helper/utils/exception.js
@@ -1,0 +1,11 @@
+class ResponseException {
+  constructor(message) {
+    this.message = message;
+  }
+
+  toString() {
+    return (this.message);
+  }
+}
+
+module.exports = { ResponseException };
