+++ src/services/user_service.js
@@ -1,15 +1,20 @@
 const cryptoRandomString = require('crypto-random-string');
+const bcrypt = require('bcrypt');
+const jwt = require('jsonwebtoken');
 const db = require('../models');
+
+const { ResponseException } = require('../helper/utils/exception');
 
 const saltRounds = process.env.SALT_ROUNDS ? parseInt(process.env.SALT_ROUNDS, 10) : 10;
 const saltLength = process.env.SALT_ROUNDS ? parseInt(process.env.SALT_LENGTH, 10) : 14;
+const jwtPrivateKey = process.env.PRIVATE_KEY_JWT || '!bE8JX7!owd!W67&XEU9kw2W';
 
 const UserService = {
   async signUp(userDTO) {
     const { username, password } = userDTO;
     const user = await db.Users.findOne({ where: { username } });
     if (user !== null) {
-      return buildRes(res, false, 'User exist');
+      throw new ResponseException('User exist');
     }
     try {
       const salt = cryptoRandomString({ length: saltLength });
@@ -19,9 +24,42 @@ const UserService = {
         password: hashedPassword,
         salt,
       });
-      return buildRes(res, true, {});
+      return { username };
     } catch (e) {
-      return buildRes(res, false, e);
+      throw new ResponseException(e.toString());
+    }
+  },
+
+  async login(userDTO) {
+    const { username, password } = userDTO;
+    try {
+      const user = await db.Users.findOne({
+        where: { username },
+      });
+
+      if (user === null) {
+        throw new ResponseException('Invalid login info');
+      }
+      const encryptedTruePassword = user.password;
+      const { salt } = user;
+
+      const isPasswordCorrect = await bcrypt.compare(
+        password + salt,
+        encryptedTruePassword,
+      );
+
+      if (isPasswordCorrect) {
+        const token = jwt.sign(
+          { username, id: user.userId, r: user.role },
+          jwtPrivateKey,
+          { expiresIn: 8640000 },
+        ); // 100 days
+        return token;
+      }
+      throw new ResponseException('Invalid login info');
+    } catch (e) {
+      console.error(e);
+      throw new ResponseException(e.toString());
     }
   },
 };
